pyproject.toml
[project]
name = "storybook-pipeline"
version = "0.1.0"
description = "Multi-agent storybook ad pipeline (LangGraph + Gemini + MongoDB)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "langgraph>=0.2.0",
  "langchain-core>=0.3.0",
  "google-genai>=1.0.0",
  "pydantic>=2.0.0",
  "pydantic-settings>=2.0.0",
  "motor>=3.6.0",
  "python-multipart>=0.0.12",
  "aiofiles>=24.0.0",
  "boto3>=1.35.0",
  "chromadb>=0.5.0",
  "openai>=1.40.0",
]

[project.optional-dependencies]
dev = ["httpx", "ruff"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
